requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-mistralai==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.128
markdown-it-py==3.0.0
MarkupSafe==3.0.0
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.9.1
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
playwright==1.47.0
prompt_toolkit==3.0.47
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
streamlit==1.39.0
tenacity==8.5.0
tokenizers==0.20.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.12.1
